+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   * `event/active-element.js`
   * `event/interaction-type-listener.js`
   * `event/shadow-focus.js`
+  * `focus/source.js`
 * adapting to general dis/engage with options API:
   * `fix-browser/pointer-focus-children.js`
   * `fix-browser/pointer-focus-input.js`